+++ src/query.ts
@@ -4,15 +4,15 @@ import { maybe, Maybe } from '@fluss/core/maybe';
  * Select element from parent node. By default selects from **document**.
  * Safe variant of `document.querySelector` method.
  */
-export function querySelector<T extends keyof HTMLElementTagNameMap>(
+export function query<T extends keyof HTMLElementTagNameMap>(
   selector: T,
   parent?: ParentNode | Maybe<ParentNode> | null
 ): Maybe<HTMLElementTagNameMap[T]>;
-export function querySelector<T extends Element>(
+export function query<T extends Element>(
   selector: string,
   parent?: ParentNode | Maybe<ParentNode> | null
 ): Maybe<T | null>;
-export function querySelector<T extends Element>(
+export function query<T extends Element>(
   selector: string,
   parent?: ParentNode | Maybe<ParentNode> | null
 ): Maybe<T | null> {