+++ test/remove_attribute.test.js
@@ -1,11 +1,11 @@
-import { hasAttribute, querySelector, removeAttribute } from '../build';
+import { hasAttribute, query, removeAttribute } from '../build';
 
 describe('removeAttribute', () => {
   test('removeAttribute removes attribute', () => {
     document.body.innerHTML = '<p class="p"></p>';
 
     expect(
-      querySelector('p')
+      query('p')
         .map((el) => {
           removeAttribute(el, 'class');
           return el;
@@ -16,12 +16,12 @@ describe('removeAttribute', () => {
   });
 
   test('removeAttribute accepts Maybe object.', () => {
-    expect(() => removeAttribute(querySelector('p'), 'class')).not.toThrow();
+    expect(() => removeAttribute(query('p'), 'class')).not.toThrow();
   });
 
   test('removeAttribute does nothing if attribute does not exists.', () => {
     expect(
-      querySelector('p')
+      query('p')
         .map((el) => {
           removeAttribute(el, 'id');
           return el;