+++ test/set_attribute.test.js
@@ -1,21 +1,21 @@
-import { setAttribute, querySelector, getAttribute } from '../build';
+import { setAttribute, query, getAttribute } from '../build';
 
 describe('setAttribute', () => {
   test('setAttribute sets attribute of element', () => {
     document.body.innerHTML = '<div></div>';
 
-    querySelector('div').map((el) => setAttribute(el, 'class', 'el'));
+    query('div').map((el) => setAttribute(el, 'class', 'el'));
     expect(
-      querySelector('div')
+      query('div')
         .map((el) => el.getAttribute('class'))
         .extract()
     ).toBe('el');
   });
 
   test('setAttribute accepts Maybe object.', () => {
-    setAttribute(querySelector('div'), 'id', 'outer');
+    setAttribute(query('div'), 'id', 'outer');
 
-    expect(getAttribute(querySelector('div'), 'id').isJust()).toBe(true);
-    expect(getAttribute(querySelector('div'), 'id').extract()).toBe('outer');
+    expect(getAttribute(query('div'), 'id').isJust()).toBe(true);
+    expect(getAttribute(query('div'), 'id').extract()).toBe('outer');
   });
 });